fix(api): ffixed msp2 serial api
